Dashboard/requirements.txt
python-dotenv
psycopg2-binary
confluent-kafka
Flask
plotly
dash
dash-bootstrap-components
boto3